requirements.txt
# Core dependencies
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
python-dotenv>=0.19.0

# Optional: For visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For progress bars
tqdm>=4.65.0
